+++ src/Northwind/SqlServer/Tables/01. Categories.sql
@@ -1,6 +1,6 @@
 ﻿DROP TABLE IF EXISTS Categories;
 CREATE TABLE Categories (
-    CategoryId      INT             IDENTITY(1, 1) NOT NULL,
+    CategoryId      SMALLINT        IDENTITY(1, 1) NOT NULL,
     CategoryName    NVARCHAR(15)    NOT NULL,
     "Description"   NTEXT           NULL,
     Picture         IMAGE           NULL,
